fix: resolve CI failures — concurrency errors, file size gate
- KeyboardShortcuts: use @preconcurrency import CoreFoundation and
  nonisolated(unsafe) for properties accessed across isolation boundaries;
  capture locals before Thread closure to avoid @MainActor property access
  from @Sendable context (fixes warnings-as-errors build)
- Extract SubredditRow + SubredditDropDelegate from ChannelsView into
  SubredditRow.swift (376→176 + 202 lines)
- Move CaptureFormResult to Models/CaptureFormResult.swift (304→295 lines)
- Regenerate xcodeproj to include new files
